Correct logic of "final sert"
Previously the while loop checked the word length and alphabetization in the same conditional; resulting in breaking the loop at *either* finding a longer word or finding one that comes later in the alphabet. 

This was causing words like "goldfish" to come before words like "pony" because goldfish comes first alphabetically, even though pony is a shorter word.