Corrected "final sert" while loop conditional
Changed the conditional of the final sert while loop to compare the length of the new word with the length of temp->next.
Was previously comparing the lengths with the current temp word, which caused the loop to automatically increment if the new word was shorter than temp. The problem with that is that the new word may come next in the list, if it is alphabetically first of the next group by word length.